+++ traits/Utilities.php
@@ -54,4 +54,52 @@ trait Utilities {
             return "page error";
         }
     }
+
+    // cart, product, products, order-process fájlokban van
+    function offer($weekoffer, $price)
+    {
+        $val = 0.9;
+        $_SESSION["week_offer"] = $val;
+        $t = getdate();
+        // $t["wday"] -> csere vmi számra teszthez
+        $_SESSION["today"] = $t["wday"];
+
+        if ($_SESSION["today"] === 0 || $_SESSION["today"] === 6) {
+            if($weekoffer === "1") {
+                return [$price, $price * $_SESSION["week_offer"]];
+            }
+            else {
+                return [$price];
+            }
+        } else {
+            return [$price];
+        }
+    }
+
+    function navOffer() {
+        $t = getdate();
+        // $t["wday"] -> csere vmi számra teszthez
+        $_SESSION["today"] = $t["wday"];
+
+        if ($_SESSION["today"] === 0 || $_SESSION["today"] === 6) {
+            return 'Ne hagyd ki a kiváló lehetőségeket, hiszen már elindultak hétvégi akcióink!';
+        } else {
+            return 'Hétvégi különleges ajánlatainkban a kiemelt termékeink akár 10%-kal kezdvezőbb áron elérhetők!';
+        }
+    }
+
+    function subTotal($week, $ertek)
+    {
+        $val = 0.9;
+        $_SESSION["week_offer"] = $val;
+        $t = getdate();
+        // $t["wday"] -> csere vmi számra teszthez
+        $_SESSION["today"] = $t["wday"];
+
+        if ($_SESSION["today"] === 0 || $_SESSION["today"] === 6) {
+            return $week === "1" ? ($ertek *= $_SESSION["week_offer"]) : $ertek;
+        } else {
+            return $ertek = 1;
+        }
+    }
 }
