+++ views/products.php
@@ -67,7 +67,7 @@
               <h6 class="name-color">Termék ára</h6>
               <div class="card-text text-color">';
                 // érdemes lenne külön fájlba rakni mint pl cart.php és product.php
-                $weekArr = $week->offer($termek["week_offer"], $termek["price"]);
+                $weekArr = $page->offer($termek["week_offer"], $termek["price"]);
                 for($i = 0; $i < count($weekArr); $i++) {
                   if(count($weekArr) === 1) {
                     echo '<p>'.$weekArr[$i].' Ft</p>';
